+++ lib/couchbase-orm/json_schema.rb
